python-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic-settings==2.4.0
